Remove splat, testing, and packaging related files
Remove test infrastructure and packaging infrastructure that is no
longer necessary.

Test infrastructure is unnecessary because the SPL is now well-tested by
the ZFS test suite, which is run by the automated builder.   The splat
utility and module are exclusively used for testing.

Packaging infrastructure will be made redundant by merging the SPL code
into the ZFS project.

This commit will not be merged into master upstream.  It is used to
verify what code need not be copied into the ZFS repo.

This commit successfully builds from a clean repo.  ZFS 0.7.0 builds
against it successfully.